fix: Improve error handling, limit volume of downloaded data and improve UX
The app downloads the 100 most recent locations posted by students.
The app gracefully handles a failure to download student locations.
The Student Locations Tabbed View has a logout button in the upper left corner of the navigation bar.
The logout button causes the Student Locations Tabbed View to dismiss, and logs out of the current session.
An activity indicator is displayed during geocoding, and returns to normal state on completion.
The app shows a placemark on a map via the geocoded response. The app zooms the map into an appropriate region.